working on package list screen
