+++ src/components/PackageRelated/Packages.tsx
@@ -8,109 +8,20 @@ import {
   chakra,
   Spacer,
   HStack,
-  IconButton,
-  useToast,
-  Text,
-  useDisclosure,
 } from '@chakra-ui/react';
-import React, { useState, useEffect, Suspense } from 'react';
-import { RiInstallLine, RiCheckLine } from 'react-icons/ri';
+import React, { useState, useEffect } from 'react';
 import {
-  useRecoilState, useRecoilCallback, useRecoilValue,
+  useRecoilState,
 } from 'recoil';
 import _ from 'lodash';
 import { invoke } from '@tauri-apps/api/tauri';
 import {
-  Category, getPackages, getPackageStatus, installPackage, packageState,
+  packageState,
 } from '../../stores/PackageStore';
+import PackageUnit from './PackageUnit';
 
 interface PackageProps {
 }
-interface PackageInfoProps {
-  isInstalled:boolean,
-  catId:string,
-  pkId:string,
-  pkgName:string,
-}
-
-const PackageInfo: React.FC<PackageInfoProps> = (props) => {
-  const toast = useToast();
-  const packageStatusUpdate = useRecoilCallback(({ snapshot, set }) => async (
-    catId:string,
-    pkId:string,
-  ) => {
-    const packages = await snapshot.getPromise(packageState);
-    const cats = _.cloneDeep(packages);
-    const catIndex = cats.findIndex((element) => element.id === catId);
-    const pkIndex = cats[catIndex].packages.findIndex((element) => element.id === pkId);
-    const updatedPk = cats[catIndex].packages[pkIndex];
-    const packageStatus = await snapshot.getPromise(getPackageStatus(updatedPk.pkg));
-    updatedPk.isInstalled = packageStatus === 'true';
-    set(packageState, (prev) => prev.map((item, index) => {
-      if (index !== catIndex) {
-        return item;
-      }
-      return {
-        ...item,
-        packages: item.packages.map((pk, index2) => (pkIndex !== index2 ? pk : updatedPk)),
-      };
-    }));
-  }, []);
-  const installPackageWithName = useRecoilCallback(({ snapshot }) => async (
-    catId:string,
-    pkId:string,
-    pkgName:string,
-  ) => {
-    const result:string = await snapshot.getPromise(installPackage(pkgName));
-
-    const desc = result.replaceAll('"', '').replaceAll('\\u{a0}', ' ').split('\\n').map((item, index) => (
-      <span>
-        {item}
-        <br />
-      </span>
-    ));
-    const colDesc = (
-      <>
-        <Text noOfLines={[1, 2, 3]}>
-          {desc}
-        </Text>
-      </>
-    );
-    packageStatusUpdate(catId, pkId);
-    if (result.toUpperCase().includes('ERROR')) {
-      toast({
-        title: `Installing ${pkgName}`,
-        description: desc,
-        status: 'error',
-        duration: 9000,
-        isClosable: true,
-        position: 'bottom-right',
-      });
-    } else {
-      toast({
-        title: `Installing ${pkgName}`,
-        description: colDesc,
-        status: 'success',
-        duration: 9000,
-        isClosable: true,
-        position: 'bottom-right',
-      });
-    }
-  });
-  const {
-    isInstalled, catId, pkId, pkgName,
-  } = props;
-  return (
-    <div>
-      {isInstalled ? (
-        <IconButton aria-label="installed" disabled icon={<RiCheckLine />} colorScheme="gray" variant="solid" />
-      ) : (
-        <IconButton aria-label="install" icon={<RiInstallLine />} onClick={() => installPackageWithName(catId, pkId, pkgName)} colorScheme="green" variant="solid" />
-      )}
-
-    </div>
-  );
-};
 
 const PackagesList: React.FC<PackageProps> = (props) => {
   const [packageSt, setPackageSt] = useRecoilState(packageState);
@@ -157,7 +68,7 @@ const PackagesList: React.FC<PackageProps> = (props) => {
         </chakra.h3>
         <Spacer />
         {pkStatus ? (
-          <PackageInfo
+          <PackageUnit
             catId={props.catId}
             pkId={props.uniqueId}
             pkgName={props.pkg}
